lambda/requirements.txt
requests==1.1.0
python-firebase
python-dotenv